Primary merge issues were:
- add_preference_to_tm ported twice (I picked the version most like
  the current CSoar implementation)
- EOL issues so had to use "git merge master -s recursive -Xignore-space-at-eol"
- unclear if there are any ordering issues betweeen epmem and wma activation
  code

+++ jsoar-core/src/main/java/org/jsoar/kernel/io/quick/MemoryNode.java
@@ -97,7 +97,7 @@ class MemoryNode
     {
         if(value instanceof Number)
         {
-            return ((Number) value).intValue();
+            return ((Number) value).longValue();
         }
         else if(value != null)
         {
